refactor(summary): extract media details
