+++ projects/client/src/lib/sections/summary/components/details/MediaDetails.spec.ts
