+++ projects/client/src/lib/sections/summary/components/details/MediaDetailsProps.ts
