+++ projects/client/src/lib/sections/summary/components/details/_internal/DetailsGrid.svelte
