+++ projects/client/src/lib/sections/summary/components/details/_internal/getMediaCost.ts
@@ -1,6 +1,6 @@
 import { languageTag } from '$lib/features/i18n/index.ts';
+import type { StreamOnDemand } from '$lib/requests/models/StreamingServiceOptions.ts';
 import { toHumanCurrency } from '$lib/utils/formatting/currency/toHumanCurrency.ts';
-import type { StreamOnDemand } from '../../../../../requests/models/StreamingServiceOptions.ts';
 
 export function getMediaCost(onDemandService: StreamOnDemand) {
   const price = onDemandService.prices.rent ?? onDemandService.prices.purchase;
